// src/app/api/model/address.ts
/**
 * WebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Address { 
    id?: string;
    city?: string;
    region?: string;
    postalCode?: string;
    country?: string;
    street?: string;
    apartment?: string;
}